refc: restructure files and locations
